Remove eboot XMC support
This is not needed as the esp8266 always runs eboot with the
flash access speed at 20MHz, and at that speed there is no need
for special handling of the XMC chips.

This commit retains the code that adds a verify stage to the
eboot copy phase as it may be useful.